Extract Billing::MaterializeProgressRenderer to shared lib
Both `billing plans materialize` and `billing catalog sync` rendered
MaterializePlans streaming events with duplicated describe/format
logic. Move the renderer to lib/ as a first-class Billing utility
with a configurable indent parameter so standalone (indent=2) and
composite (indent=4) contexts get the correct nesting.

Also fix catalog sync to surface materialization errors in the final
banner instead of always reporting success — `step_materialize` now
returns false when errors occurred so the closing message distinguishes
"complete" from "finished with errors".
